Refactored main classes, added tests

+++ src/main/java/com/example/coderecognizer/data/repository/ProductCodeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductCodeRepository extends JpaRepository<ProductCode, Long> {
+public interface ProductCodeRepository extends JpaRepository<ProductCode, Integer> {
 }
